backend/requirements.txt
Flask
python.dotenv
flask-restful
Flask_SQLAlchemy